Moving types to their own files
