+++ src/Kinetic/ObservableObject.cs
@@ -163,83 +163,3 @@ public abstract class ObservableObject
         }
     }
 }
-
-[DebuggerDisplay("Get()")]
-public readonly struct Property<T>
-{
-    internal readonly ObservableObject Owner;
-    internal readonly IntPtr Offset;
-
-    internal Property(ObservableObject owner, IntPtr offset) =>
-        (Owner, Offset) = (owner, offset);
-
-    public T Get() => Owner.Get<T>(Offset);
-
-    public void Set(T value) => Owner.Set(Offset, value);
-
-    public IObservable<T> Changed => Owner.EnsureObservableFor<T>(Offset);
-
-    public static implicit operator T(Property<T> property) =>
-        property.Get();
-
-    public static implicit operator ReadOnlyProperty<T>(Property<T> property) =>
-        new ReadOnlyProperty<T>(property.Owner, property.Offset);
-}
-
-[DebuggerDisplay("Get()")]
-public readonly struct ReadOnlyProperty<T>
-{
-    internal readonly ObservableObject Owner;
-    internal readonly IntPtr Offset;
-
-    internal ReadOnlyProperty(ObservableObject owner, IntPtr offset) =>
-        (Owner, Offset) = (owner, offset);
-
-    public T Get() => Owner.Get<T>(Offset);
-
-    public IObservable<T> Changed => Owner.EnsureObservableFor<T>(Offset);
-
-    public static implicit operator T(ReadOnlyProperty<T> property) =>
-        property.Get();
-}
-
-internal abstract class PropertyObservable
-{
-    internal readonly ObservableObject Owner;
-    internal readonly PropertyObservable? Next;
-    internal readonly IntPtr Offset;
-
-    internal uint Version;
-
-    protected PropertyObservable(ObservableObject owner, IntPtr offset, PropertyObservable? next) =>
-        (Owner, Offset, Next) = (owner, offset, next);
-
-    public abstract void Changed();
-}
-
-internal sealed class PropertyObservable<T> : PropertyObservable, IObservableInternal<T>
-{
-    private ObservableSubscriptions<T> _subscriptions;
-
-    public PropertyObservable(ObservableObject owner, IntPtr offset, PropertyObservable? next) :
-        base(owner, offset, next)
-    { }
-
-    public override void Changed() =>
-        Changed(Owner.Get<T>(Offset));
-
-    public void Changed(T value) =>
-        _subscriptions.OnNext(value);
-
-    public IDisposable Subscribe(IObserver<T> observer)
-    {
-        observer.OnNext(Owner.Get<T>(Offset));
-        return _subscriptions.Subscribe(this, observer);
-    }
-
-    public void Subscribe(ObservableSubscription<T> subscription) =>
-        _subscriptions.Subscribe(this, subscription);
-
-    public void Unsubscribe(ObservableSubscription<T> subscription) =>
-        _subscriptions.Unsubscribe(subscription);
-}
